Change dollar_amount type from Float to Integer in ChoresUser model

+++ models/user_data/chores/chores.py
@@ -5,7 +5,7 @@ class ChoresUser(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), unique=True, nullable=False)
-    dollar_amount = db.Column(db.Float, nullable=False, default=0.0)
+    dollar_amount = db.Column(db.Integer, nullable=False, default=0)
     household_id = db.Column(db.Integer, db.ForeignKey('households.id'), nullable=True)
     household_admin = db.Column(db.Boolean, default=False)
